EES-6244 Create ReplacementPlanService
