+++ src/GovUk.Education.ExploreEducationStatistics.Admin/Controllers/Api/DataReplacementController.cs
@@ -1,8 +1,6 @@
 using System;
-using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
-using GovUk.Education.ExploreEducationStatistics.Admin.Models;
 using GovUk.Education.ExploreEducationStatistics.Admin.Requests;
 using GovUk.Education.ExploreEducationStatistics.Admin.Services.Interfaces;
 using GovUk.Education.ExploreEducationStatistics.Admin.ViewModels;
@@ -17,7 +15,7 @@ namespace GovUk.Education.ExploreEducationStatistics.Admin.Controllers.Api;
 [Authorize]
 [ApiController]
 public class DataReplacementController(
-    IReplacementService replacementService,
+    IReplacementPlanService replacementPlanService,
     IReplacementBatchService replacementBatchService)
     : ControllerBase
 {
@@ -26,7 +24,7 @@ public class DataReplacementController(
         Guid originalFileId,
         CancellationToken cancellationToken = default)
     {
-        return await replacementService.GetReplacementPlan(
+        return await replacementPlanService.GetReplacementPlan(
                 releaseVersionId: releaseVersionId,
                 originalFileId: originalFileId,
                 cancellationToken: cancellationToken
