+++ src/GovUk.Education.ExploreEducationStatistics.Admin/Services/Interfaces/IReplacementService.cs
@@ -2,7 +2,6 @@
 using System;
 using System.Threading;
 using System.Threading.Tasks;
-using GovUk.Education.ExploreEducationStatistics.Admin.ViewModels;
 using GovUk.Education.ExploreEducationStatistics.Common.Model;
 using Microsoft.AspNetCore.Mvc;
 
@@ -10,11 +9,6 @@ namespace GovUk.Education.ExploreEducationStatistics.Admin.Services.Interfaces
 {
     public interface IReplacementService
     {
-        Task<Either<ActionResult, DataReplacementPlanViewModel>> GetReplacementPlan(
-            Guid releaseVersionId,
-            Guid originalFileId,
-            CancellationToken cancellationToken = default);
-
         Task<Either<ActionResult, Unit>> Replace(
             Guid releaseVersionId,
             Guid originalFileId,
