+++ src/GovUk.Education.ExploreEducationStatistics.Admin/Startup.cs
@@ -481,6 +481,7 @@ namespace GovUk.Education.ExploreEducationStatistics.Admin
             services.AddTransient<IRelatedInformationService, RelatedInformationService>();
             services.AddTransient<IReplacementService, ReplacementService>();
             services.AddTransient<IReplacementBatchService, ReplacementBatchService>();
+            services.AddTransient<IReplacementPlanService, ReplacementPlanService>();
             services.AddTransient<IUserRoleService, UserRoleService>();
             services.AddTransient<IUserReleaseRoleService, UserReleaseRoleService>();
             services.AddTransient<IUserPublicationRoleRepository, UserPublicationRoleRepository>();
